file: pages/story 이동 및 구조변경
